+++ models/dataloader.py
@@ -29,8 +29,8 @@ class Dataset_3Dfront(Dataset):
 
     def __getitem__(self, index):
         return (
-            torch.tensor(self.layout_dataset[index], dtype=torch.float32),
-            torch.tensor(self.mask_dataset[index], dtype=torch.float32),
+            torch.tensor(self.layout_dataset[index], dtype=torch.bfloat16),
+            torch.tensor(self.mask_dataset[index], dtype=torch.bfloat16),
         )
 
     def __len__(self):
